Restructured component folder
